requirements.txt
fastapi
uvicorn
pydantic
pyyaml
numpy
requests
huggingface_hub
